fix: profileImage null 처리 메서드 적용, 구독자 수 처리 수정, 커서 조건 추가

+++ src/main/java/com/fade/subscribe/dto/response/FindSubscriberResponse.java
@@ -5,7 +5,7 @@ import java.util.List;
 public record FindSubscriberResponse(
         List<FindSubscriberItemResponse> subscribers,
         Long nextCursor,
-        Integer totalSubscribers
+        Long totalSubscribers
 ) {
     public record FindSubscriberItemResponse(
         Long id,
